Test 8 and modify old tests
The old tests were designed to show that the correct number of letters would be returned even if the number was slightly bigger than it had to be. Now this causes issues because the smaller numbers are represented by letters as well. This was also a large step because the logic that makes 1000s, 900s, 500s, and 400s work is the same as the logic that makes 100s, 90s, 50s, and 40s work.